+++ src/Microsoft.IdentityModel.Tokens.Saml/Saml/Experimental/SamlValidationError.cs
