+++ src/Microsoft.IdentityModel.Tokens.Saml/Saml2/Experimental/Saml2ValidationError.cs
