+++ src/Microsoft.IdentityModel.Tokens/Experimental/SecurityTokenException.Experimental.cs
