+++ src/Microsoft.IdentityModel.Tokens/Experimental/TokenHandler.Internal.cs
