+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Exceptions.cd
