+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/Details/AlgorithmValidationError.cs
