+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/Details/IssuerValidationError.cs
