+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/Details/SignatureValidationError.cs
