+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/Details/TokenTypeValidationError.cs
