+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/ValidatedIssuer.cs
