+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/ValidatedLifetime.cs
