+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/ValidatedSigningKeyLifetime.cs
