+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/ValidatedToken.cs
