+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Results/ValidatedTokenType.cs
