+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/ValidationDelegates.cd
