+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/ValidationFailureType.cs
