+++ src/Microsoft.IdentityModel.Tokens/Experimental/Validation/Validators.TokenType.cs
