+++ src/Microsoft.IdentityModel.Validators/Experimental/AadIssuerValidator.Internal.cs
