+++ src/Microsoft.IdentityModel.Xml/Experimental/XmlValidationException.Experimental.cs
